updating login from sso
