+++ Backend/ManagerLayer/ManagerLayer.csproj
@@ -52,11 +52,11 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Logging\LoggingManager.cs" />
-    <Compile Include="Login\UserManager.cs" />
+    <Compile Include="User\UserManager.cs" />
     <Compile Include="PointManagement\PointManager.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="AccessControl\AuthorizationManager.cs" />
-    <Compile Include="Login\UserLoginManager.cs" />
+    <Compile Include="SSOUtility\UserLoginManager.cs" />
     <Compile Include="UserManagement\UserManagementManager.cs" />
   </ItemGroup>
   <ItemGroup />
