+++ Backend/ServiceLayer/Services/ExceptionService.cs
@@ -78,6 +78,13 @@ namespace ServiceLayer.Services
             public UserNotFoundException(string message, Exception innerException) : base(message, innerException) { }
         }
 
+        public class UserAlreadyExistsException : Exception
+        {
+            public UserAlreadyExistsException() { }
+            public UserAlreadyExistsException(string message) : base(message) { }
+            public UserAlreadyExistsException(string message, Exception innerException) : base(message, innerException) { }
+        }
+
         public class InvalidHeaderException : Exception
         {
             public InvalidHeaderException() { }
